keeper-service/requirements.txt
firebase-admin==6.9.0
Flask==3.0.2
gunicorn==22.0.0
python-dotenv==1.0.1